server/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
ldap3==2.9.1
Markdown==3.7
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
